MAINT: retain the field IO inside the domain due to parallel concerns
